+++ src/Editor.tsx
@@ -1,12 +1,8 @@
 import { styled } from "@macaron-css/solid";
 import { createMemo, createSignal, For, onMount, Show } from "solid-js";
-import { Clear } from "./Clear";
-import { Header } from "./Header";
-import { Help } from "./Help";
 import { useRegionContext } from "./Region";
 import { useAppStore } from "./store";
 import { type Point as PointId, type Quad } from "./store/editor";
-import { Upload } from "./Upload";
 import { v } from "./vector";
 
 type Point = { x: number; y: number };
@@ -31,25 +27,6 @@ const Canvas = styled("svg", {
   },
 });
 
-const Toolbar = styled("div", {
-  base: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    backgroundColor: "whitesmoke",
-    padding: "5px 10px",
-    width: "100%",
-    whiteSpace: "nowrap",
-  },
-});
-
-const Buttons = styled("div", {
-  base: {
-    display: "flex",
-    gap: "5px",
-  },
-});
-
 export function Editor() {
   const [store] = useAppStore().file;
 
@@ -67,30 +44,6 @@ export function Editor() {
         <DrawingBoard imageRef={imageRef()!} />
       </Show>
     </Container>
-  );
-}
-
-export function EditorToolbar() {
-  const [store] = useAppStore().file;
-  const width = () => store.image.naturalWidth;
-  const height = () => store.image.naturalHeight;
-
-  // TODO remove
-  const move = useRegionContext();
-
-  return (
-    <Toolbar>
-      <div>
-        Size: {width()} x {height()} Current: {move.current().x},{" "}
-        {move.current().y} Origin: {move.origin().x}, {move.origin().y}
-      </div>
-      <Buttons>
-        <Clear />
-        <Upload />
-        <Help />
-        <Header />
-      </Buttons>
-    </Toolbar>
   );
 }
