+++ src/toolbar/Clear.tsx
